Organised all the Checkin Code into one Class
