+++ Big Neon/View/ScannerModeView/ShowGuestListView.swift
@@ -37,7 +37,6 @@ public class ShowGuestListView: UIView {
     public override init(frame: CGRect) {
         super.init(frame: frame)
         self.backgroundColor = UIColor.white
-        self.loadingView.startAnimating()
         self.configureView()
     }
